requirements.txt
Flask
numpy
scipy
matplotlib
gunicorn
google-generativeai
pandas
scikit-learn
gunicorn

